added authorization to todos endpoints

+++ routers/todos.py
@@ -3,8 +3,9 @@ from sqlalchemy.orm import Session
 from fastapi import APIRouter, Depends, HTTPException, Path
 from pydantic import BaseModel, Field
 from models import Todos
-from database import  SessionLocal
+from database import SessionLocal
 from starlette import status
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -18,6 +19,7 @@ def get_db():
 
 
 dependency_injection = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -28,29 +30,39 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/")
-async def read_all(db: dependency_injection):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: dependency_injection):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authenticated Failed!")
+    return db.query(Todos).filter(Todos.owner == user.get('id')).all()
 
 
 @router.get("/todos/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: dependency_injection, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: dependency_injection, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authenticated Failed!")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get("id")).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found!")
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: dependency_injection, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.dict())
+async def create_todo(user: user_dependency, db: dependency_injection, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authenticated Failed!")
+
+    todo_model = Todos(**todo_request.dict(), owner_id=user.get("id"))
 
     db.add(todo_model)
     db.commit()
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: dependency_injection, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: dependency_injection, todo_request: TodoRequest,
+                      todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authenticated Failed!")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found!")
 
@@ -64,9 +76,11 @@ async def update_todo(db: dependency_injection, todo_request: TodoRequest, todo_
 
 
 @router.delete("todos/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: dependency_injection, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id)
+async def delete_todo(user: user_dependency, db: dependency_injection, todo_id: int = Path(gt=0)):
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed!")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found!")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get("id")).delete()
     db.commit()
